Update Wikibase site id and group for test2wiki and testwiki
Use the "default" for site id, which is $wgDBname and
put testwiki and test2wiki in the special sites group
so we can add links to them from test wikidata.

Also remove the disabling of propagate moves/deletions to repo.

The repo is test.wikidata (repoDatabase and changesDatabase settings)
for test.wikipedia and test2.wikipedia.

Bug: T94416
Change-Id: I18df3171ba204cb584ebcc895396dc14064ba4e0