Update Wikibase site id and group for test2wiki and testwiki
Use the "default" for site id, which is $wgDBname and
put testwiki and test2wiki in the special sites group
so we can add links to them from test wikidata.

also remove the disabling of propagate moves/deletions
to repo. the repo is test.wikidata so this should be
okay now.

Bug: T94416
Change-Id: I18df3171ba204cb584ebcc895396dc14064ba4e0