Make json_encode() options backward-compatible with PHP 5.3
We should really upgrade the few remaining hosts that are still on PHP 5.3.
In the meantime, pass a literal value of 448 as $options for json_encode(),
equivalent to JSON_PRETTY_PRINT | JSON_UNESCAPED_SLASHES | JSON_UNESCAPED_UNICODE,
but doesn't break on PHP 5.3. This way we don't have to remember to flip it
back once tin is upgraded.

Change-Id: Iabaaf0e50a8f2e00c22cff6033c24daf29d9e78c