Update patch set 1
Patch Set 1:

As detailed in the ticket beta is more complicated. Because it uses the same wiki id for the config as www.wikidata.org so divering those is worse than divering test. from www. .

Or do you think it is fine to also enable it on www.wikidata.org?

test.wikidata.org is a client that has arbitrary access enabled.

Lydia said the request is just to get it enabled somewhere together with wikidata to test. If you prefer it to be enabled somehwere else, I'm open to suggestions.

Patch-set: 1
Reviewer: JanZerebecki <640@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0