Update patch set 4
Patch Set 4:

I wonder if it makes sense to do this more slowly/piecemeal. It might be safer.

In my estimation, the newer and more specific lists (e.g., group1.dblist and gettingstarted-with-category-suggestions.dblist) are likely easier to move to a different location than the older and more established lists such as all.dblist.

Patch-set: 4
Label: Verified=0