Update patch set 1
Patch Set 1:

Are the browsers clever enough not to re-fetch the fonts again when browsing different wiki? There is the font-name parameter, but I doubt they use it because that would allow rogue websites to change appearance of other websites. In my browser fonts are now re-loaded for every wiki.

Is this an acceptable regression in bandwidth usage and in length of flash of unstyled content while the fonts are loaded?

Patch-set: 1
Reviewer: Nikerabbit <93@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0