Update patch set 1
Patch Set 1:

That's what baffles me the most. This is a php project. I'm better equipped to write PHP than other languages. And most of all, the repository I'm committing this to is all PHP, and this is added next to other php command line tools. It's entirely appropriate, IMO

Patch-set: 1