Update patch set 1
Patch Set 1:

I want to reiterate, I'm not against writing this in <insert-any-language-but-csh>, I'm against reinventing the wheel. And I feel like we're doing that a bit here, but I may be missing some details: I've used quilt for years and always felt like it's a godsend to manage patches you have to cherry-pick onto an upstream repository - it's Debian's tool of choice for a reason.

But again, your choice.

Patch-set: 1