Update patch set 1
Patch Set 1:

Thing is, this simple script I wrote in php does the trick, and honestly using nothing at all would be easier than arguing about it which is why I've let this sit without really fighting for it any further until Greg weighed in yesterday.

I was about half way through addressing Bryans code review as of this morning.  I'm not particularly fond of this solution and really it's not a big deal to continue with essentially no tools. 90% of the benefit was gained by simply keeping the patches in /srv/patches, using version control on that directory mainly just provides a bit more accountability by logging changes.

Patch-set: 1