Update patch set 1
Patch Set 1:

a) The current solution sucks. I think we can all agree on that. b) we now have something that works, and means we're not going to accidentally drop security patches c) there's nothing stopping us from moving to some already-written tool in the future, but IMO the advantage of not dropping security patches (which has happened enough times in the past) is worth going with something that already works instead of a perfect solution.

Patch-set: 1