Update patch set 1
Patch Set 1:

What's actually wrong now is that the version of Elasticsearch we have deployed seems to be missing some dependencies needed for lucene expression language. Its not clear to me why we didn't catch it in development, but part of the reason is that we mostly tested with the wikimedia-extra plugin fully enabled. Those tests passed fine. This patch enables the plugin.

Patch-set: 1
Reviewer: Manybubbles <873@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0