Disable Parsoid cache updates
Now that all active public wikis are enabled in RESTBase* and the major clients
are switched over, we can stop updating the Parsoid caches proactively. This
will halve the number of parse requests from template updates, and indirectly
lower the load on the API cluster and related infrastructure.

*: This should not be deployed before I10c8c37013d7, which actually enables
the remaining public & active wikis.

Bug: T93291
Change-Id: Ibea7fccce0f29992adc10eb5b2ba362335c447e3