Hide "prefershttps" preference from everywhere
As the HTTPS-only migration is underway, this setting will make
progressively no sense to all wikis and is extra confusing. Noone new
should be unsetting this (as it won't work soon) and while there's some
point with existing users setting it, it's going to be done for all of
them soon anyway.

Bug: T102245
Change-Id: I915897b1189a7ab868aa1f05a75ba5f87687855c