Remove obsolete 'https -> http' rewrite for IRC notifications
This was added when HTTPS was not the default to avoid the urls in
IRC notifications from varying on the protocol of the incoming request
(e.g. when users could opt-in to HTTPS).

Since HTTPS is the default everywhere, this regex no longer matches.
Even if it did work, we shouldn't emit HTTP urls anywhere as they just
get redirected to HTTPS with a 301.

Bug: T122933
Change-Id: I152aa757678f6934b91ea5caf4f1a218f275d471