wgCanonicalServer->https for new HTTPS domains
This is intended to cover the 7 sets of language domains which
have been converted to HTTPS only in the past few days.  The set
of domains in varnish is the $lang domainname prefix for all
projects for these 7 languages currently.  I pulled the set of
applicable projects from grepping all.dblist for the language
prefixes.

Change-Id: I3694489ba14b312a167773ecc522a091e5833efc