Pull out unnecessary wikitech settings, move some into CommonSettings
There's no reason for wgPasswordSenderName to be different on wikitech
specifically, wgCookieDomain seems like it would be the default assumed
by browsers anyway, wgMainCacheType is already the given value everywhere,
same for wgSessionsInObjectCache. We should not enable wgShowExceptionDetails
in production so remove that comment. Also remove comments about old
config, that's unneeded now. If, for whatever terrible reason, we decide
to have another wmgUseClusterFileBackend = false, that can get InsantCommons
as well.

Bug: T75939
Change-Id: Idfcbe35bf8020fe45235c836278626ea05a746cf