Update patch set 6
Patch Set 6: Code-Review-1

If we want to use the approach "Echo maintenance scripts need to know they should not run if Echo is disabled" (I think this is a little hacky and would prefer dblist expressions, but I'm not going to block this approach), we need to do so on all maintenance scripts.

There are 5 Echo maint scripts, and only one has this check.

Patch-set: 6
Label: Code-Review=-1