Update patch set 6
Patch Set 6:

The only reason that script gets such a check is because it needs to be run systematically by a cron job... No other script is, and the same goes for pretty much every other extension maintenance script where the extension is not installed on all Wikimedia sites. Echo is not special, please remove your -1.

Patch-set: 6
Label: Verified=0