Never use wgServer/wgCanonicalServer values from production in labs
Setting enwiki's wgCanonicalServer to https caused the value
of wgCanonicalServer in labs to be overridden by the one in production,
so emails sent from beta enwiki had links pointing to production enwiki.

Use a minus sign for wgServer and wgCanonicalServer to ensure that
beta will never use server URLs from the production config.

Change-Id: I46cd9add4cf99c158b597c8939e12505b21ba46b