Add nonglobal.dblist, for labswiki
labswiki was added to fishbowl.dblist to stop confusing CentralAuth
(I3d9a3f48), but since fishbowl.dblist is intended for
restricted-editing wikis this has the side effect of also disabling
various anti-spam measures that aren't really necessary when editing is
restricted to only trusted accounts. Since labswiki allows public
account creation, this is bad.

This patch creates a new dblist for wikis that shouldn't be global but
are otherwise publicly editable, and moves labswiki to that list
instead.

Change-Id: I5540b13cd2a2b8ae9408be86a7437481c4c85b60