Update patch set 2
Patch Set 2: Code-Review+1

Agree it shouldn't be a problem in most cases.

I remember hoo was having problem with a Wikidata widget, and mobile was having issues uploading directly to commons (using the api/CORS, iirc). But we can start with this as is, and if those teams are still having issues, we can potentially add those wikis later on.

Patch-set: 2
Reviewer: CSteipp <204@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1