More high-resolution logos for Chinese Wikipedia
This patch submits the other variant logo for the
Chinese Wikipedia to /static. It includes the regular
logo as well as the 1.5x and 2x versions for higher
resolution displays.

Both logos have been generated from the official wiki
logo and optimized with optipng.

The logo is copyrighted by the Wikimedia Foundation and
released under the CC-BY-SA; also trademarked by the WMF.

See Phabricator task for details.

Bug: T102852
Change-Id: I78966ced2e6f451e463dd6b3c43b61a4da2c90d2