Update patch set 1
Patch Set 1: Code-Review-1

Blerg. Yeah this won't work (at least not deterministically) due to the way in which the logging system is wired up internally. Once LoggerFactory::getInstance() has been called the array will have been passed to ObjectFactory and then on to MonologSpi. At that point modifying the original array will trigger PHP's copy-on-write semantics and modify the global but detach it from the copy that is used to create logger instances.

So possible fixes (in order of effort) are:
* log the ldap channel all of the time as a normal logging channel that goes to fluorine
* extend the custom logging config generation here in mw-config to have a syntax for local log output and log all the time to silver
* Add additional accessors to the LoggerSpi interface and MonologSpi that let you extend the logging configuration post SPI creation

Patch-set: 1
Label: Code-Review=-1