static: Add foundation logo (with hidpi variants)
The old foundationwiki.png was based on
https://wikimediafoundation.org/wiki/File:Wikimediafoundation-logo.png

This is an outdated version that rendered the font slightly off.
The official logo is
https://commons.wikimedia.org/wiki/File:Wmf_logo_vert_pms.svg
which has the text converted to shapes (irregardless of locally
installed fonts). As referenced from:
https://wikimediafoundation.org/wiki/Wikimedia_trademarks

The RGB and square variant is
https://commons.wikimedia.org/wiki/Wikimedia_Foundation_RGB_logo_with_text.svg

Using thumbs of that, with sizes 135px (as before), and 204px/270px
(matching enwiki, enwiki-1.5x and enwiki-2x).

Run through optipng, otherwise unmodified.

Change-Id: Ia0fdf45e2974f7f7340aaa5b8d29de9abe952d81