favicon/touch icon proxy: Fix broken redirect to internal (http)
Turns out favicon.php mistakenly used PROTO_INTERNAL to expand
$wgFavicon. In the past the favicon was a full url pointing to
bits.wikimedia.org. This was recently changed to a file in the
domain-relative /static/ directory. And it finally broke when we
enforced HTTPS since this proxy script doesn't connect to internal
servers but public ones, and it doesn't handle redirects.

Bug: T104538
Change-Id: I5b1da82fc7846d5308de7aeb6f064094de79f783