Update patch set 3
Patch Set 3:

Code looks good, but I can't find in the commit nor the referenced task anything about why Redis. In case this goes side-ways and/or for when in the future we want to change this, some indication of why Redis (and/or why not what we currently use) - would be useful for historical record.

Patch-set: 3