Set a different wmgContentTranslationDefaultSourceLanguage for English
This is a simplistic solution for Bug T105327 - the English Wikipedia
shows an unlikely source language by default.

The comprehensive solution for bug T105327 is in
I8eb35c688525451e4692b160830ab7ccf27851ce
but this should be merged as well, because the default
source for English cannot be English anyway.
For now setting it to Spanish, which is the most popular
translation target language by number of translators.
This can be adjusted later.

Bug: T105327
Change-Id: I177835ffcc37d04ad2806f1fd66fb5165d56b894