Set a different wmgContentTranslationDefaultSourceLanguage for English
This is a simplistic solution for Bug T105327 - the English Wikipedia
shows an unlikely source language by default.

The default source for English cannot be English anyway,
so for now setting it to Spanish, which is the most popular
translation target language by number of translators.
This can be adjusted later.

The really comprehensive solution for the bug would be
to take the default source language from the ULS common languages,
and that will be a separate patch for the extension.

Bug: T105327
Change-Id: I177835ffcc37d04ad2806f1fd66fb5165d56b894