Prevent race condition when writing settings to cache
fopen() with 'w' can lead to a race condition where the file is
half-assembled (either by truncation or the write being incomplete).

Avoid this by truncating after the file is opened and locked instead
of at fopen() time and by releasing the lock only after the file is
completely done.

Bug: T103744
Change-Id: Ibdfee686ce5c499cea56a24ea717686f912cf51b