Get rid of default=wikipedia assumptions in config
Stuff should break when you try to do that. (This commit attempts to prevent
existing cases from changing however.)

Also get rid of 'wikispecies' group entries on beta that won't do anything
because there is only one wikispecies site and it doesn't have a
wikispecies.dblist.

I have not actually tested this yet and it doesn't fix the default logo.

Bug: T104088
Change-Id: Idfc4bebc9871f5e78947ae997f5a8a6a91006b41