Update patch set 3
Patch Set 3: Code-Review-1

Please announce ahead of time through wikitech, wikitech-ambassadors and Tech News to allow the community time to track down any tools and dependencies thereof that may depend on the url format.

Remember that irc.wikimedia.org is basically exclusively consumed through regexes. This will break stuff. Let's at least try to limit that damage for tools that are maintained.

I'd give it a month or so.

Patch-set: 3