Enable Flow on all wikis, except private and a couple special wikis
This includes wikis without VisualEditor.

Private wikis are those that the public can't even read.

The "special" wikis are loginwiki and votewiki.

This does not actually put Flow on any pages, unless/until
someone explicitly enables it on particular page/namespace.

Also, add a sanity check that Parsoid is enabled.

Bug: T106562
Change-Id: I12af04743752469b7370fd9aabeed21a7ef7cc8c