Bump cache epoche for wikidata
there were some changes (I683358e) in the site links
ui that need this. A "collapse" toggle was added to
each site link section.

without bumping cache epoch, then the toggle only
appears on pages not currently in parser cache or
pages that have been edited / purged. perhaps that
is ok enough, but think it's somewhat weird and 
confusing to have the toggle only on some pages.

Change-Id: I1784b05fd7c79a143ad27844130649a6729c091a