Double $wgMemoryLimit (330 => 660)
HHVM's memory utilisation is substantially better than PHP 5.3, and our
application server fleet has substantially more RAM than it used to. We
often get fatals when calling HH\xenon_get_data(), which means that there
is potentially a class of requests that are both memory-intensive and
utterly invisible to xenon profiling. So let's double the limit and see what
happens.

Change-Id: I3c6217f06783ae1630bf9aa3d7b7eabc14f2bafa