Use relative URLs to construct load.php requests
As far as I can determine, the use of absolute URLs was mandated by the
requirement to load ResourceLoader assets from the bits.wikimedia.org domain.
When bits was deprecated, we unthinkingly made the new URLs absolute, although
this was no longer a requirement, and indeed prevents the mobile site from
loading modules via the mobile domain.

Bug: T106966
Change-Id: I6665bf317f7f3293afb078e733aa7c69fd404fe9