logging: Only send info and higher to logstash by default
Change default configuration for logging so that debug level events are
not sent to Logstash. The MediaWiki usage of PSR-3 is such that debug
level must be explicitly used by the PHP code and is intended only for
low level logging events that are too verbose to be useful in
operational logging.

Change-Id: I9dc2328f57a32348683d8b7e2148a625c0023b45