Allow import from any WMF project to any other
It's been a long time coming, but here it is.

The old wgImportSources configuration has been completely removed, as there
is no need for it. In its place, we have a hook function
"wmfImportSources" which lazy loads an importation list for the whole WMF
cluster.

Also adding a missing ) to the import user agent.

IMPORTANT: DO NOT MERGE until the following REQUIRED dependencies are merged:
* I68989203e367e7ea515e4ae2222c330b264a7cb1
* Ice9a19cb6dfe53ae72aa71353d0553ee9338f233

Bug: T17583
Change-Id: Ib325df7c76837e1936b320d2a2441f10405090cd