Allow import from any Labs/Beta Cluster project to any other
(Beta Cluster only for now.)

For labs, we have a hook function
"wmfImportSources" which lazy loads an importation list for the whole WMF
cluster.

Old config is still used for production.

Also adding a missing ) to the import user agent.

IMPORTANT: DO NOT MERGE until the following REQUIRED dependency is deployed on all sites:
* I68989203e367e7ea515e4ae2222c330b264a7cb1

Bug: T17583
Change-Id: Ib325df7c76837e1936b320d2a2441f10405090cd