Update patch set 2
Patch Set 2:

> Was a proper fix for this ever applied?

Yes, I think that the changes in Iba6f115 will allow this patch to be rewritten in a way that actually updates the config. I forgot all about this patch at some point. I'll update it to use MonologSpi::mergeConfig() and we can retest. Thanks for the reminder poke.

Patch-set: 2