Update patch set 2
Patch Set 2:

Not a dev, but I think it needs to be rebased? Also (very minor), it seems that the practice is to add simply "T..." instead of "Bug T..." (I don't really care much). Otherwise changes looks good for me, however I do agree that we might discuss if we can do this for at least all WMF projects as major projects such as enwiki, dewiki and frwiki have requested that this be restricted to this usergroups. Maybe we should open a specific for that topic or open a thread at m:Wikimedia Forum or m:RFC. Best regards.

Patch-set: 2
Reviewer: MarcoAurelio <787@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0