Update patch set 1
Patch Set 1:

Yes, it was Jimmy's idea in the first place, but Jimmy is not blocking alteration of that setting. According to Greg Maxwell he was in favour of reverting it in 2007: http://article.gmane.org/gmane.science.linguistics.wikipedia.english/84182 

My position in that discussion was to just handle it like a normal shell request: http://thread.gmane.org/gmane.science.linguistics.wikipedia.english/84133

I don't think such a request was ever made at the time, since the community was split on the issue. In 2008 there was T18642, a conservative proposal allowing users to create articles as AFC subpages. In 2011, there was a community proposal to restrict page creation even further, requiring autoconfirmed status (T32208), but this was rejected despite majority support, mostly due to opposition from staff. For example, Brandon wrote "We have buckets and buckets of hard data that says this change will harm the encyclopedia."

So at this point, the reason for the setting is the opinions of the users.

Patch-set: 1
Label: Code-Review=0