Replicate browser test config for QuickSurveys
Tests are currently failing on beta labs due to invalid configuration.
This copy pastes the config changes from the test given that there is
currently no better way to do this.

Potentially we could include the LocalSettings.php in the test folder
but that seems like it could cause us headaches down the road so this is
deemed much safer.

Bug: T112204
Change-Id: I23e4385ff18423e0a7cc11a6d9d9f50151681296