Change default AbuseFilter IP block duration to not indefinite
Indefinite IP blocks are discouraged on WMF projects. This is a default
for WMF projects so any project who allows the AbuseFilter to block as
action do not indefinite block IP addresses, but just blocks them for one
week. If the IP is problematic then it'll be reblocked by the AbuseFilter
or the block can be extended by local administrators. This change does not
affect local wikis which have opted for different block durations as
configured by $wgAbuseFilterAnonBlockDuration and
$wgAbuseFilterBlockDuration.

Bug: T113164
Change-Id: Ic6ed54071b1e7a1405a4635b9d13d646c8d2cff6