Refactor monolog handling for kafka logs
The new kafka logging from CirrusSearch will want to only log to
kafka and not dump these binary logs to fluorine.  This patch
refactors so handlers can enable udp2log, logstash and/or kafka
as desired.

This additionally adds a new test case that validates all
channels are configured in a valid manner.

Bug: T103505
Change-Id: Ib1a474318091844964ee0dc5133ea354e3d09275