Maintain existing api.log format when adding context
The api request logs are moving to a structured format, but
we don't want to log that structured format to fluorine. The
file size would just be too much in addition to the existing
textual log lines.

Patch adjusts to allow api to specify the formatter to use,
and makes a formatter that doesn't include the context in
it's output

Bug: T108618
Change-Id: I715e62c4ec0340b12e01ba5a5c84887b620ce722