Update patch set 1
Patch Set 1:

I've started pt-heartbeat on es2 and es3 (es1 is ro), and provided extra grants needed.

It should be ready for small tests now, but I will not retire the -1 until it is properly puppetized and monitored on icinga (which I will be doing soon)- it is not correctly setup on MariaDB10 masters.

Patch-set: 1