Update patch set 2
Patch Set 2: Code-Review-1

I am only -1'ing as a formality so I don't forget why this is held up.  We noticed these failures are already hitting graphite now (which we could alert on) but are also including false negatives which would make an alert on this not useful yet.  So we can't rely it for communicating time ranges for repopulation. 

https://graphite.wikimedia.org/render?from=-2weeks&until=now&width=500&height=400&target=logstash.rate.mediawiki.CirrusSearchChangeFailed.WARNING.sum&drawNullAsZero=true&lineMode=staircase&title=logstash.rate.mediawiki.CirrusSearchChangeFailed.WARNING.sum

I think erik is refactoring to compensate

Patch-set: 2
Label: Code-Review=-1