Update patch set 2
Patch Set 2:

I think that this will affect only jobs that are replayed because the cluster is frozen. I don't know yet how the job queue handles failed jobs (limited number of attempts? timeouts?)
So in the case of a lab failure I'm not sure we have a cirrus config var that controls the number of retries/timeout.

Patch-set: 2