Update patch set 2
Patch Set 2:

I apologise for that - it is just moving chunks of text about, changing the bug ID's to Txxxxx and one fix to some broken code. I don't know how to make this easier - I could break it down into several, but then there is still an issue of moving chunks about in the page.

Patch-set: 2
Label: Verified=0