Update patch set 2
Patch Set 2:

Well currently jenkins wont run the extension-unittests-generic test because the extension dosent follow the naming of the repo in the files. It Does Timeline.php not timeline.php as the repo is called. I created a timeline.php file and EasyTimline.php file is the main php file we would want to use since that is what the author is calling it.

But there isent really a transition period since you carnt have both lowercase and uppercase of the same file.

Patch-set: 2
Reviewer: Paladox <1665@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0