Log messages sent to the 'warning' channel
When a php notice or warning occurs it gets sent to this 'warning'
channel via MWDebug::warning. The same logs end up in hhvm.log and
the 'hhvm' channel in logstash, but neither contains information
such as the wiki that triggered the error or the url that was running.

Adding this channel will end up duplicating those logs, but with
more information useful for debugging.

Change-Id: Ic032fd12239a19ec1325586d6db934891dcc1938