Remove override of $wgEchoDefaultNotificationTypes
This was used in 2012/2013 to roll out Echo in a controlled
fashion, then left equal to the default value in Echo, and
got out of sync with Echo over time.

The practical effect of this change is to no longer send
emails for welcome notifications, which ae3b1f29 in Echo
(August 2014) attempted to do but failed to because of this
override.

Bug: T113367
Change-Id: Ie335d5284105fb4a22080103fa5ff76a4dd141b7