Update patch set 2
Patch Set 2:

Not really a setting can be called what ever they like as long as they follow the mediawiki convention on how they are set out. The settings are following the repo name. Whereas the files are following the name it is actually called in the extension credits and on the mediawiki extension page.

Add backwords support to files is easier then doing it with settings.

Patch-set: 2