Update patch set 3
Patch Set 3:

I respect your opinion, Krenair (always have), but in this specific case we need to move ahead with the decision by Chris Steipp and myself. We weighed the issues and this is the best short term solution for all. We believe that this limited additional enablement is OK as no further enablements are planned to happen any time soon without addressing our collective concerns.

Patch-set: 3