Add "composer test" command to lint files and run tests
You need to fetch the development dependencies to run tests, by running
"composer update --dev" first.

This is kind of messy, because we have part of the vendor directory
checked-in, but development dependencies should not be. Ideally the
multiversion code could be split out into its own respository or
something.

Bug: T85947
Change-Id: I3fa9ac84b9882a0579cd3f81a7adc088792f9f4d