Update patch set 3
Patch Set 3:

Yes. I've made that assumption based on the fact this survey = https://phabricator.wikimedia.org/T113443 = needs to be run next week.

I've asked for clarification on your behalf, if this is really a crucial detail but eventually I imagine we'll want to be able to get in a position we can run arbitrary surveys on all our projects to learn from our readers but this is not what is needed right now.

Patch-set: 3
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0