Enable banners extension on mobile web beta only (enwiki)
MobileFrontend has some code that predates the WikidataPageBanner.
It enables banners only in the beta mode of the mobile skin.
(Banners will be blacklisted in other skins)

This turns the extension on there to allow us to run some tests
and launch a survey to gather community feedback about the
feature and allows us to remove that code in MobileFrontend.

It will not render banners in any of the blacklisted skins so business
as usual there.

Bug: T101108
Change-Id: If0539c53c6ab83aa39ef09ac76ccbf8b64584c5b