Update patch set 4
Patch Set 4: Code-Review-1

(1 comment)

Oops, minor oversight on my part. The extension-list changes need to roll out with the initial 1.27-wmf.6 deploy on Tuesday so that l10n caches get built by scap. The extension can't be enabled on enwiki until the wmf.6 branch gets there on Thursday however so that will need to be a separate patch.

This initial patch should probably enable the extension on testwiki for testing in prod and setup a simple survey for that testing.

Patch-set: 4
Label: Code-Review=-1