Monolog: Disable microsecond timestamps on all loggers
Follow up to I8ff63f2 that adds the useMicrosecondTimestamps(false) call
to several logger definitions that were missed. The '@default' logger
template is probably the most important as this logger definition is
used for all of the log events that are eventually sent to /dev/null on
the production cluster and is the most used logger definition.

Bug: T116550
Change-Id: I198d9f0f2a5537f0a3c85613d4bbd4311effaa50