Monolog: Disable microsecond timestamps on all loggers
Follow up to I8ff63f2 that adds the useMicrosecondTimestamps(false) call
to several logger definitions that were missed. The '@default' logger
template is probably the most important as this logger definition is
used for all of the log events that are eventually sent to /dev/null on
the production cluster and is the most used logger definition.

Consolidates the post-construction setter calls as
$wmgMonologLoggerCalls so they can easily be amended in the future when
Monolog 2.0 changes the forced tiemzone setting to a per-Logger setting
rather than a global setting.

Bug: T116550
Change-Id: I198d9f0f2a5537f0a3c85613d4bbd4311effaa50