Switch dispatcher subscriptionLookupMode to subscriptions only
subscriptions only mode is the default setting, so we can
remove this setting entirely from here.

all client wikis have usage and subscription tracking,
so this is safe to do now and should be done.

also remove the useLegacyChangesSubscription setting in
the labs setting, since this is the default and should
match with the subscription mode setting.

Bug: T112245
Change-Id: I052de06a63cb114f244089c496df295e8bad9570