Update patch set 1
Patch Set 1:

Technically, looks good to me.

There is support from Commons community for this.

Yet, I wonder if whitelist a whole TLD when we don't want to whitelist *.wmflabs.org makes sense. According Csteipp, the goal of the whitelist is to avoid Wikimedia servers to flood external servers and to avoid attacks from them.

Patch-set: 1