Update patch set 1
Patch Set 1: Code-Review+1

To deploy without making the serializers complain I think we need to deploy the new field starting at the end of the pipeline (camus) which will apply the empty default. The deploy this which adds the empty field when serializing in php.  Then finally we can deploy a CirrusSearch patch that starts sending the field.

Patch-set: 1
Label: Code-Review=+1