Update patch set 3
Patch Set 3:

Updated this patch to very specifically only measure a count of number of logins that would be affected by T104370-T104373 (so it logs based on group membership, not rights, since that is how password policies are enforced). I'll put in a followup patch for non-SUL wikis when we decide to update those too, but I don't have the capacity to tackle that right now.

Patch-set: 3
Label: Verified=0