Update patch set 1
Patch Set 1:

Since we have accounts logging in with passwords that are too short, I think this is appropriate so that we can determine the impact updating the policy will have. We should remove this once we feel we have a reasonable idea of the impact.

Patch-set: 1
Reviewer: CSteipp <204@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0