Update patch set 1
Patch Set 1: Code-Review-1

Commit message should be rewritten.

The title should get a rough summary of what is being done, be concise but precise. For example 'Add group <name of the group> to enwikiversity'. a line in the 50 chars is perfect, in the 60 the upper limit 

Then, each line should have < 72 chars, ideally < 60.

Use the thumbrule A paragraph = an idea.

Here you have two ideas:
* what rights are in the group
* who can add/remove the group

Patch-set: 1
Label: Code-Review=-1