Enable subpages on custom aliases from 112 to 119
We've enabled in the past subpages for custom namespaces.
A bunch of them were forgotten for wikis wih a lot of namespaces.

We add a litle more of namespaces in the configuration to help to
cope with the namespaces inflation, from 112 to 119.

For example, T119207 requires a new namespace with subpages,
and it will receive an ID > 110.

Change-Id: I3998a8e29c2b28e01ac17014035bb5d7e94f1345