Increase $wgCopyUploadTimeout to 90 seconds (from default 25)
25 seconds may be a sane default for most http requests, but the
copy upload feature is used for uploading big files, and often
sees timeouts. Obviously, commons will probably try to use this
feature to upload gb big files, and that's still going to timeout,
but 90 seconds is a much more reasonable timeout for normal sized
files. I imagine we wouldn't want to increase it too much beyond
that (although we could go a bit), as we want to cut the http
request with enough time left over to actually upload the file.

The php execution limit should be 120 seconds for the types of
requests in question ($wgTransactionalTimeLimit), so we should
definitely hit the copy upload timeout before hitting that timeout.
I believe the timeout on the varnish side is 180 seconds, and
I believe the timeout for the url-downloader.wikimedia.org proxy is
5 minutes, so $wgCopyUploadTimeout should still be the controlling
timeout for these types of requests.

Bug: T118887
Change-Id: Ie33ae3b6a6cb757fe1b2a644e4361ea029789440