Lower redis connection timeout from 1s to 0.5s
The connection timeout was introduced in I6b99685b (and set to 1s) and then
raised to 2s in Ie23cc06f, presumably to make MediaWiki more accomodating of
increased load on the job queue servers. Since we raised the number of redis
backends in active use from 2 to 9 in the last week, this is no longer an
issue, so it is safer for MediaWiki to assume that a server that is slow to
respond is down or unreachable.

I think even 0.5s is way too high, but let's start with that.

Change-Id: I5d1d5c680b7fc7a1e748b8a0ef986099c9b7fa9e