Update patch set 1
Patch Set 1:

It would work on testwiki because testwiki responses are not cached in varnish. It would make more sense to target test2wiki, but then you'd simply run into T43451 again. In other words: depending on whether you target testwiki or test2wiki, this experiment would either confirm that the feature works as intended when there is no cache layer, or that it does not work as intended when there is one. I don't see much value in that, but I don't mind merging this if you want to see it first-hand.

Patch-set: 1
Reviewer: Ori.livneh <255@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0