Enabled ogg opus support for TimedMediaHandler
There are running plans to enable Ogg Encapsulation
for the Opus Audio Codec in the TimedMediaHandler extension
by default.

Before to upgrade the extension code, we should first test
in real world the feature, for example, on Wikimedia Commons
to get feedback.

This change so enables it for Wikimedia sites.

Change-Id: Id1e765c8357570888e4c3e926f9bad3e76aaa85c