Enable cluster-wide import setup on the Beta Cluster
(Beta Cluster only for now.)

For labs, we have a hook function "wmfImportSources"
which lazy loads an importation list for the whole WMF
cluster. The setup is actually intended for production,
and is being tested on labs to make sure it works
correctly. Some of the wikis included as import sources
do not exist on the Beta Cluster.

Old config is still used for production.

Bug: T17583
Change-Id: Ib325df7c76837e1936b320d2a2441f10405090cd