Enable cross-wiki upload A/B test in additional languages
The following 9 languages now have a complete translation of the A/B
test messages (and the rest of the dialog): de, fr, he, it, mk, nap,
nl, pl, zh-hans. Around 17% of the cross-wiki uploads so far were from
projects in these languages. (Around 43% were in English.)

Enabling the test on all of them, except for *wikimedia ones. It is
unfortunately impossible to do it by content language, so we have to
list all the projects.

Also add enwiktionary, which I somehow missed in fea3ff57.

https://www.mediawiki.org/wiki/Multimedia/December_2015_cross-wiki_upload_A/B_test

Bug: T120867
Change-Id: I12a7ab2a15cd33c2f9f9093eb42d6b9b2a64e169