Update patch set 6
Patch Set 6: Code-Review-1

Normaly the group is named "rollbacker", the right is "rollback", so please change this to default, otherwise translations will not match, so users at other languages can't see a translation of the word rollbacker in their language. Default is "rollbacker" (for example see line 8794).

Patch-set: 6
Label: Code-Review=-1