Adjust cirrus titlesuggest index shard counts
Now that we've built the initial indices we have a better idea
of what size they are. Adjust down the number of shards to aim
for 150MB < x < 500MB per shard. This is still a pretty wide
range and a bit arbitrary, but seems reasonable. We can add
replicas at any time via the es rest api if the query load needs
to be better spread out.

Bug: T124332
Change-Id: I17193364274d55fc7c56c8772270e3de29ab9ec4