Basic "Identifiers" statement section config for Wikidata
Note that only a selection of a few heavily used string property ids is
listed (including VIAF, ISBN and GND, to name the most relevant). The
full list would be some hundred property ids. This is not worth the
trouble, in my opinion. We start with a basic set for testing. Later,
the community decides which properties are identifiers and which not.
These will be changed to have the "external-id" type. When this
happened we will switch the configuration as described in T123112.

Bug: T123112
Change-Id: Ie4e28a90c8221b5fc4b49bdfcabb7cf278c18853