Update patch set 1
Patch Set 1:

Not sure what you are referring to. It's used in millions of Wikipedia articles as one of the most critical and relevant identifiers, linking to the German national library. I do not find anything about it being obsolete or something in it's article, see https://de.wikipedia.org/wiki/Gemeinsame_Normdatei .

But that's besides the point anyway. This will actually *lower* the visibility of these statements because they move down to a second section.

Patch-set: 1
Label: Verified=0