Update patch set 2
Patch Set 2:

if the config is going to be same for test.wikidata, beta and wikidata, then please define it just once in Wikibase.php

(better yet, why can't we have a default in wikibase? what did Lydia not like about that? that would make it easier for developers, allow vagrant to be the same, etc.)

Patch-set: 2
Label: Verified=0