Centralize and add rights and grants in preparation for grants moving into core
To make things easier to track in the future, let's move all the bits
setting $wgAvailableRights into one place in CommonSettings.php.

Then, in preparation for moving of grants into core, we need to
duplicate all the settings of $wgMWOAuthGrantPermissions into
$wgGrantPermissions. Once 1.27.0-wmf.11 is everywhere, we can then
remove the setting of $wgMWOAuthGrantPermissions.

And while we're at it, let's add some stuff that was missing into
$wgGrantPermissions:
* FlaggedRevs 'movestable' should probably go with editprotected
* The 'templateeditor' and 'editeditorprotected' should definitely go
  with editprotected.
* TitleBlacklist's 'tboverride' should also go with editprotected
* TitleBlacklist's 'tboverride-account' should probably go with
  'createaccount'.

Change-Id: I7d9f82c6216eed50e8de6a72923e1756c2a81671