Establish `group2` and `exempt` dblists
Having an `exempt.dblist` will accommodate blacklisting for cases where
we need to temporarily hold certain wikis back, the immediate use case
being for `labswiki` and `labstestwiki`.

Using an explicit `group2.dblist` instead of `all.dblist` will allow us
to implement this blacklist for the final train deployment. On-wiki
documentation for train (and perhaps other) deployment should be updated
to reflect this.

Change-Id: I9a0f9d86cc006af23540b2ecccfe466c63a91b51